hello_flask/requirements.txt
flask==3.0.0
google-genai==0.2.2